refactor: move AssertFinalExpIsOne to sw_bls12381
