+++ std/algebra/emulated/fields_bls12381/e12_pairing.go
@@ -157,68 +157,3 @@ func (e *Ext12) MulBy02368(a *E12, c0, c1 *E2) *E12 {
 		A11: *d11,
 	}
 }
-
-// AssertFinalExponentiationIsOne checks that a Miller function output x lies in the
-// same equivalence class as the reduced pairing. This replaces the final
-// exponentiation step in-circuit.
-// The method is inspired from [On Proving Pairings] paper by A. Novakovic and
-// L. Eagen, and is based on a personal communication with A. Novakovic.
-//
-// [On Proving Pairings]: https://eprint.iacr.org/2024/640.pdf
-func (e Ext12) AssertFinalExponentiationIsOne(x *E12) {
-	tower := e.ToTower(x)
-
-	res, err := e.fp.NewHint(finalExpHint, 24, tower[0], tower[1], tower[2], tower[3], tower[4], tower[5], tower[6], tower[7], tower[8], tower[9], tower[10], tower[11])
-	if err != nil {
-		// err is non-nil only for invalid number of inputs
-		panic(err)
-	}
-
-	if err != nil {
-		// err is non-nil only for invalid number of inputs
-		panic(err)
-	}
-
-	residueWitness := e.FromTower([12]*baseEl{res[0], res[1], res[2], res[3], res[4], res[5], res[6], res[7], res[8], res[9], res[10], res[11]})
-	// constrain cubicNonResiduePower to be in Fp6
-	// that is: a100=a101=a110=a111=a120=a121=0
-	// or
-	//     A0  =  a000 - a001
-	//     A1  =  0
-	//     A2  =  a010 - a011
-	//     A3  =  0
-	//     A4  =  a020 - a021
-	//     A5  =  0
-	//     A6  =  a001
-	//     A7  =  0
-	//     A8  =  a011
-	//     A9  =  0
-	//     A10 =  a021
-	//     A11 =  0
-	scalingFactor := E12{
-		A0:  *e.fp.Sub(res[12], res[13]),
-		A1:  *e.fp.Zero(),
-		A2:  *e.fp.Sub(res[14], res[15]),
-		A3:  *e.fp.Zero(),
-		A4:  *e.fp.Sub(res[16], res[17]),
-		A5:  *e.fp.Zero(),
-		A6:  *res[13],
-		A7:  *e.fp.Zero(),
-		A8:  *res[15],
-		A9:  *e.fp.Zero(),
-		A10: *res[17],
-		A11: *e.fp.Zero(),
-	}
-
-	// Check that  x * scalingFactor == residueWitness^(q-u)
-	// where u=-0xd201000000010000 is the BLS12-381 seed,
-	// and residueWitness, scalingFactor from the hint.
-	t0 := e.Frobenius(residueWitness)
-	// exponentiation by -u
-	t1 := e.ExptNeg(residueWitness)
-	t0 = e.Mul(t0, t1)
-
-	t1 = e.Mul(x, &scalingFactor)
-
-	e.AssertIsEqual(t0, t1)
-}
